Created BaseView and NormalView.
The idea I'm going for is that any given sequence on the display will be
a "view".  Views get called on each frame to determine what needs to go
on the display (besides color changes).  It's that simple.

This creates BaseView (the base class, with helper methods) and
NormalView (the "idle" state of the project, displaying current run
stats).