Made NormalView handle transitions to Service[Menu|Credit]View.
Now that functionally all other views are handling buttons in some way,
shape, or form (either ServiceMenuView navigating its menus or
SimpleAnimationView (and derivatives) handling the Menu button to cancel
the animation), it doesn't make much sense that entering the menus (or
ServiceCreditView) from NormalView is handled by a special case in the
main loop.

To that end, NormalView now handles those transitions itself, and the
handle_buttons method (derived from BaseView) no longer returns True or
False, it returns an instance of BaseView (if switching to a new view)
or None (if not).

Note that this is a little bit flaky in the button-handling part of
things; new ButtonDatas are only sent if there's change in the state of
ANY button, and the logic in, say, NormalView will therefore trigger if,
for instance, Back is held down when exiting ServiceMenuView and then
Menu is pressed.  That's honestly not a particularly realistic use
scenario, and I doubt it's that important, but maybe I should change the
button handling logic so that it only returns the CHANGES in each button
(up, down, no change).