Fixed Omega Shift transitions on startup.
Before, the logic I had meant that if the whole program was started
during Omega Shift, it would always start with the Omega Shift
transition animation.  This isn't right; we shouldn't start with an
animation right off the bat.  Now it'll wait until at least one fetch
(WITH a definitive answer on Omega Shiftness) has completed before
displaying the animation.