Made shift transitions not show up right after boot.
At boot, the Pi might not have the correct date and time set up until it
gets its first time sync.  That time sync might not happen for a while,
as it needs to find a wifi connection first, to the point where this
program may have already started, waiting at YDIJ, with the wrong shift
color (as it's judging that based on the time at startup).  When that
gets updated with the first time sync, the color changes, but since it's
already in the main loop (just without NormalView yet), it also adds the
shift transition to whatever the "real" shift now is, meaning it'll go
straight into the animation as soon as the first fetch happens.

With all that said, this change makes it so that shift transitions
aren't put in the view queue if stats haven't been fetched yet.  This
means it'll never launch into a shift animation directly from YDIJ,
which makes more sense.